Pause function implemented. Other general module updates

+++ core/src/edu/vcu/nopreference/Objects/physics/Collision.java
@@ -9,10 +9,10 @@ import edu.vcu.nopreference.Objects.base.objectBase;
 public class Collision {
     objectBase object1; //active
     objectBase object2; //passive
-    int collidedX = 0;//-1 left, 0 none, 1 right
-    int collidedY = 0;//-1 down, 0 none, 1 up
+    float collidedX = 0;//-1 left, 0 none, 1 right
+    float collidedY = 0;//-1 down, 0 none, 1 up
     boolean collided;
-    public Collision(objectBase active, objectBase passive, int CollidedX, int CollidedY, boolean Collided)
+    public Collision(objectBase active, objectBase passive, float CollidedX, float CollidedY, boolean Collided)
     {
         object1 = active;
         object2 = passive;
@@ -29,11 +29,11 @@ public class Collision {
         return object2;
     }
 
-    public int getCollidedX() {
+    public float getCollidedX() {
         return collidedX;
     }
 
-    public int getCollidedY() {
+    public float getCollidedY() {
         return collidedY;
     }
     public boolean getCollided()
